api/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.88.0
h11==0.14.0
idna==3.4
numpy==1.24.0
Pillow==9.3.0
pydantic==1.10.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
keras==2.11.0
tensorflow==2.11.0
Keras-Preprocessing==1.1.2